CHK-7037: Remove Disused Checkout Shipping Save Event Observer (#85)
Fixes "SHIPPING_OPTION_NOT_SELECTED" error thrown by Bold Checkout
Update Order API for Digital Wallets orders placed from the Magento
Checkout. This error was thrown because the preceeding call to save the
shipping method in Magento failed due to the logic in the event observer
that makes a call to hydrate the order using the Bold Checkout API. This
hydration logic was removed by PR #71 for ticket CHK-6850, resulting in
the error "Bold update failed" being thrown by the observer because the
code it referenced no longer exists.